migrated functions out to their own class
